backend/requirements.txt
Flask==3.1.0
flask-cors==4.0.1
python-dotenv==1.0.1
appwrite>=13.4,<14
Flask
gunicorn
python-dotenv
